Expose (export) all user-facing types
Previously, only the the `EleventyFromNotionOptions` type was exported.
While this makes sense for projects that have a single descriptive type
for their options, this package defines it as a union of other types,
that are not exported. This makes the documentation terrible, because
none of the _actually_ helpful types are explained or even shown.