Correct captions in block formatters
The type for the `info.caption` property of the `BlockInfo` passed to
block formatters that have a caption was incorrect; previously, it was
defined to be of type `string`, but instead, it should be rich text.
This not only updates the types but also the default block formatters
that use captions, since they were incorrectly rendering `[object
Object]`.