Fix header rows and columns for tables
The default block formatter for tables incorrectly applied a row header
to tables that in Notion have a column header and vice versa.

The confusion here stems from the fact that Notion calls these options
"header row" and "header column" in their UI, and these correspond to
`has_column_header` and `has_row_header` properties, respectively. This
kind of makes sense; a row of headers means that these are headers for
their respective columns. However, it wasn't immediately clear to me
that this was the case, and so the behavior for the block formatter
ended up being swapped.