Complete rewrite
The previous versions, 0.3.0 and below, were not particularly
extensible. In particular, it made it hard for files to have certain
side effects, like downloading images and other assets. This rewrite
changes the API into something that looks very similar to Eleventy's
configuration style. It is easier to supply custom parsers and
formatters, and even plugins. A few plugins are shipped as built-ins: a
JSON, YAML, Markdown and HTML plugin, and a `SkipModified` plugin taking
over the reins of the similarly named option in the older versions.